requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
flask-cors==4.0.1
PyJWT==2.8.0
Werkzeug>=3.0.0

# اختبارات
pytest>=8.0.0

# رفع ملفات Excel
openpyxl>=3.1.0
